py/requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.4.0
asn1crypto==1.5.1
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.9.3
bitarray==2.5.1
bitcoin==1.1.42
bs4==0.0.1
cached-property==1.5.2
certifi==2017.7.27.1
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==6.7
coincurve==16.0.0
colorama==0.4.5
configobj==5.0.6
crypto==1.4.1
cryptography==3.4.7
cytoolz==0.11.0
dataclasses==0.8
discord.py==1.7.3
distlib==0.3.3
ecdsa==0.17.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
filelock==3.3.2
Flask==0.12.2
free-proxy==1.0.2
freezegun==0.3.9
funcsigs==1.0.2
future==0.18.2
futures==3.1.1
hexbytes==0.2.1
http-client==0.1.22
idna==2.6
idna-ssl==1.1.0
importlib-metadata==4.6.4
importlib-resources==5.2.2
ipfshttpclient==0.8.0a2
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==3.2.0
lru-dict==1.1.8
lxml==4.6.3
MarkupSafe==1.0
multiaddr==0.0.9
multidict==5.1.0
mypy-extensions==0.4.3
Naked==0.1.31
netaddr==0.8.0
numpy==1.19.5
opencv-python==4.5.5.62
parsimonious==0.8.1
Pillow==8.4.0
pipenv==2021.5.29
platformdirs==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pycurl==7.43.0.5
Pymem==1.8.5
pyOpenSSL==20.0.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyrsistent==0.18.0
pysha3==1.0.2
pystun3==1.0.0
python-dateutil==2.6.1
python-dotenv==0.19.0
pytz==2017.3
pywin32==304
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.3.1
random-user-agent==1.0.1
repoze.lru==0.7
requests==2.18.4
rlp==1.2.0
rsa==4.8
scrypt==0.8.20
shellescape==3.8.1
six==1.11.0
soupsieve==2.2.1
startrek==0.2.16
stun==0.2.3
systematic==4.8.7
theblockchainapi==0.0.41
toolz==0.11.1
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==1.4
udp==0.4.20
urllib3==1.22
uuid==1.30
varint==1.0.2
virtualenv==20.13.0
virtualenv-clone==0.5.7
web3==5.30.0
websockets==9.1
Werkzeug==0.15.3
yarl==1.6.3
zipp==3.5.0
